requirements.txt
alembic==0.7.6

flask==0.10.1
flask-migrate==1.2.0
flask-script==2.0.5
flask-sqlalchemy==2.0
